requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.4
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
whitenoise==6.6.0
cloudinary==1.39.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-email-verification==0.3.2
